third_party/WebKit/LayoutTests/media/track/track-cue-rendering.html

<!DOCTYPE html>
<title>Test that default positioned TextTrack's cues are rendered correctly.</title>
<script src="../media-file.js"></script>
<script src="../media-controls.js"></script>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<video>
    <track src="captions-webvtt/captions.vtt" kind="captions" default>
</video>
<script>
async_test(function(t) {
    var video = document.querySelector("video");
    var testTrack = document.querySelector("track");
    video.src = findMediaFile("video", "../content/test");

    var cueTextIndex = 0;
    var cueText = [ "Lorem", "ipsum", "dolor", "sit" ];
    video.onseeked = t.step_func(function() {
        assert_equals(video.currentTime, cueTextIndex + 0.5);
        assert_equals(testTrack.track.activeCues.length, 1);
        assert_equals(testTrack.track.activeCues[0].text, cueText[cueTextIndex]);

        var testCueDisplayBox = textTrackDisplayElement(video, "display");
        assert_equals(testCueDisplayBox.innerText, cueText[cueTextIndex]);
        assert_equals(2 * testCueDisplayBox.offsetLeft, video.videoWidth - testCueDisplayBox.offsetWidth)

        if (++cueTextIndex == cueText.length) {
            // Test the cue display colors and font.
            testFontSize(320, 240);
            testFontSize(640, 480);
            testFontSize(1280, 960);
            testFontSize(2560, 1440);

            assert_equals(getComputedStyle(textTrackDisplayElement(video)).fontFamily, "sans-serif");
            assert_equals(getComputedStyle(textTrackDisplayElement(video)).color, "rgb(255, 255, 255)");
            assert_equals(getComputedStyle(textTrackDisplayElement(video, "display").firstChild).backgroundColor, "rgba(0, 0, 0, 0.8)");

            t.done();
        } else {
            video.currentTime += 1;
        }
    });

    function testFontSize(width, height) {
        video.width = width;
        video.height = height;
        // Force a relayout of the document
        document.body.offsetTop;
        assert_equals(getComputedStyle(textTrackDisplayElement(video)).fontSize, parseInt(height * 0.05) + "px");
    }
    
    video.currentTime = 0.5;
});
</script>
